updated ui  + mic feature + BLoc state management
