+++ lib/constant/colors.dart
@@ -1,0 +1,4 @@
+import 'dart:ui';
+
+const backgroundColor = Color(0xff343541);
+const botBackgroundColor = Color(0xff444654);
