+++ lib/constant/secret_api_key.dart
@@ -1,0 +1,1 @@
+  const apikey = '-----PLACE YOUR API KEY HERE------';
