+++ lib/model/model.dart
@@ -1,11 +1,11 @@
 enum ChatMessageType { user, bot }
 
 class ChatMessage {
+  final String text;
+  final ChatMessageType chatMessageType;
+
   ChatMessage({
     required this.text,
     required this.chatMessageType,
   });
-
-  final String text;
-  final ChatMessageType chatMessageType;
 }
