+++ lib/views/chat_page.dart
@@ -1,0 +1,201 @@
+
+import 'package:example/views/widgets/chat_message.dart';
+import 'package:flutter/material.dart';
+import 'package:google_fonts/google_fonts.dart';
+import 'package:speech_to_text/speech_to_text.dart' as stt;
+
+import '../constant/colors.dart';
+import '../controller/generate_response.dart';
+import '../model/model.dart';
+
+
+class ChatPage extends StatefulWidget {
+  const ChatPage({Key? key}) : super(key: key);
+
+  @override
+  State<ChatPage> createState() => _ChatPageState();
+}
+
+
+
+class _ChatPageState extends State<ChatPage> {
+  final _textController = TextEditingController();
+  final _scrollController = ScrollController();
+  final List<ChatMessage> _messages = [];
+  late bool isLoading;
+  final stt.SpeechToText _speech = stt.SpeechToText();
+  bool _isListening = false;
+  TextEditingController messageTextController = TextEditingController();
+
+  @override
+  void initState() {
+    super.initState();
+    isLoading = false;
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        centerTitle: true,
+        toolbarHeight: 70,
+        title: Padding(
+          padding: const EdgeInsets.all(8.0),
+          child: Text(
+            "OpenAI ChatGPT",
+            style: GoogleFonts.michroma(),
+          ),
+        ),
+        backgroundColor: botBackgroundColor,
+      ),
+      backgroundColor: backgroundColor,
+      body: SafeArea(
+        child: Column(
+          children: [
+            Expanded(
+              child: _buildList(),
+            ),
+            Visibility(
+              visible: isLoading,
+              child: const Padding(
+                padding: EdgeInsets.all(8.0),
+                child: CircularProgressIndicator(
+                  color: Colors.white,
+                ),
+              ),
+            ),
+            Padding(
+              padding: const EdgeInsets.all(8.0),
+              child: Row(
+                children: [
+                  GestureDetector(
+                    child: const Icon(Icons.mic),
+                    onTap: _startListening,
+                  ),
+                  _buildInput(),
+                  _buildSubmit(),
+                ],
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+
+  Widget _buildSubmit() {
+    return Visibility(
+      visible: !isLoading,
+      child: IconButton(
+        color: Colors.white,
+        icon: const Icon(
+          Icons.send_rounded,
+        ),
+        onPressed: () async {
+          setState(
+            () {
+              _messages.add(
+                ChatMessage(
+                  text: _textController.text,
+                  chatMessageType: ChatMessageType.user,
+                ),
+              );
+              isLoading = true;
+            },
+          );
+          var input = _textController.text;
+          _textController.clear();
+          Future.delayed(const Duration(milliseconds: 50))
+              .then((_) => _scrollDown());
+          generateResponse(input).then((value) {
+            setState(() {
+              isLoading = false;
+              _messages.add(
+                ChatMessage(
+                  text: value,
+                  chatMessageType: ChatMessageType.bot,
+                ),
+              );
+            });
+          });
+          _textController.clear();
+          Future.delayed(const Duration(milliseconds: 50))
+              .then((_) => _scrollDown());
+        },
+      ),
+    );
+  }
+
+  Expanded _buildInput() {
+    return Expanded(
+      child: TextField(
+        textCapitalization: TextCapitalization.sentences,
+        style: const TextStyle(color: Colors.white),
+        controller: _textController,
+        decoration: const InputDecoration(
+            fillColor: botBackgroundColor,
+            filled: true,
+            border: InputBorder.none,
+            focusedBorder: InputBorder.none,
+            enabledBorder: InputBorder.none,
+            errorBorder: InputBorder.none,
+            disabledBorder: InputBorder.none,
+            hintText: 'Write something here . . .'),
+      ),
+    );
+  }
+
+  ListView _buildList() {
+    return ListView.builder(
+      controller: _scrollController,
+      itemCount: _messages.length,
+      itemBuilder: (context, index) {
+        var message = _messages[index];
+        return ChatMessageWidget(
+          text: message.text,
+          chatMessageType: message.chatMessageType,
+        );
+      },
+    );
+  }
+
+  void _scrollDown() {
+    _scrollController.animateTo(
+      _scrollController.position.maxScrollExtent,
+      duration: const Duration(milliseconds: 300),
+      curve: Curves.easeOut,
+    );
+  }
+
+  void _startListening() async {
+    if (!_isListening) {
+      bool available = await _speech.initialize(
+        onStatus: (status) {
+          // ignore: avoid_print
+          print('onStatus: $status');
+        },
+        onError: (errorNotification) {
+          // ignore: avoid_print
+          print('onError: $errorNotification');
+        },
+      );
+      if (available) {
+        setState(() {
+          _isListening = true;
+        });
+        _speech.listen(
+          onResult: (val) => setState(() {
+            _textController.text = val.recognizedWords;
+          }),
+        );
+      }
+    } else {
+      setState(() => _isListening = false);
+      _speech.stop();
+    }
+  }
+
+
+}
+
+
