+++ lib/views/widgets/chat_message.dart
@@ -1,0 +1,66 @@
+import 'package:flutter/material.dart';
+import '../../constant/colors.dart';
+import '../../model/model.dart';
+
+
+
+class ChatMessageWidget extends StatelessWidget {
+  const ChatMessageWidget(
+      {super.key, required this.text, required this.chatMessageType});
+
+  final String text;
+  final ChatMessageType chatMessageType;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      margin: const EdgeInsets.symmetric(vertical: 10.0),
+      padding: const EdgeInsets.all(16),
+      color: chatMessageType == ChatMessageType.bot
+          ? botBackgroundColor
+          : backgroundColor,
+      child: Row(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: <Widget>[
+          chatMessageType == ChatMessageType.bot
+              ? Container(
+                  margin: const EdgeInsets.only(right: 16.0),
+                  child: const CircleAvatar(
+                    backgroundColor: Color.fromRGBO(16, 163, 127, 1),
+                    child: Icon(Icons.account_box_outlined),
+                  ))
+              : Container(
+                  margin: const EdgeInsets.only(right: 16.0),
+                  child: const CircleAvatar(
+                    child: Icon(
+                      Icons.person,
+                    ),
+                  ),
+                ),
+          Expanded(
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: <Widget>[
+                Container(
+                  padding: const EdgeInsets.all(8.0),
+                  decoration: const BoxDecoration(
+                    borderRadius: BorderRadius.all(Radius.circular(8.0)),
+                  ),
+                  child: Text(
+                    text,
+                    style: Theme.of(context)
+                        .textTheme
+                        .bodyLarge
+                        ?.copyWith(color: Colors.white),
+                  ),
+                ),
+              ],
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
+
+
