+++ pubspec.yaml
@@ -36,6 +36,13 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
+  google_fonts: ^4.0.3
+  speech_to_text: ^6.1.1
+  firebase_auth: ^4.2.4
+  firebase_core: ^2.4.1
+  email_auth: ^1.0.0
+  shared_preferences: ^2.0.15
+  cloud_firestore: ^4.3.1
 
 dev_dependencies:
   flutter_test:
